requirements.txt
django==1.6.1
django-tastypie==0.11.0
